# src/ApiGateways/AdminPanel/AdminPanel.Infrastructure/Specifications/ServiceForumSwagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "County RP Forum Service API",
    "description": "The County RP Forum Service API documentation description.",
    "version": "v1"
  },
  "host": "localhost:10511",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/User": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "User_Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "apiUserDtoIn",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ApiUserDtoIn"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApiUserDtoOut"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApiUserDtoOut"
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "User_Edit",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "apiUserDtoIn",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ApiUserDtoIn"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      },
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "User_Delete",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/User/ByLogin/{login}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_GetByLogin",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "login",
            "in": "path",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApiUserDtoOut"
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
    "/User/FilterBy": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "User_FilterBy",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "Login",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "SortingFlag",
            "in": "query",
            "x-nullable": true
          },
          {
            "type": "array",
            "name": "GroupIds",
            "in": "query",
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "type": "integer",
            "name": "Count",
            "in": "query",
            "format": "int32",
            "x-nullable": true
          },
          {
            "type": "integer",
            "name": "Page",
            "in": "query",
            "format": "int32",
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/PagedFilterResultOfApiUserDtoOut"
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "ApiUserDtoOut": {
      "type": "object",
      "required": [
        "id",
        "groupId",
        "reputation",
        "posts",
        "warnings"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "login": {
          "type": "string"
        },
        "groupId": {
          "type": "integer",
          "format": "int32"
        },
        "reputation": {
          "type": "integer",
          "format": "int32"
        },
        "posts": {
          "type": "integer",
          "format": "int32"
        },
        "warnings": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ApiUserDtoIn": {
      "type": "object",
      "required": [
        "groupId",
        "reputation",
        "posts",
        "warnings"
      ],
      "properties": {
        "login": {
          "type": "string"
        },
        "groupId": {
          "type": "integer",
          "format": "int32"
        },
        "reputation": {
          "type": "integer",
          "format": "int32"
        },
        "posts": {
          "type": "integer",
          "format": "int32"
        },
        "warnings": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "PagedFilterResultOfApiUserDtoOut": {
      "type": "object",
      "required": [
        "allCount",
        "page",
        "maxPages"
      ],
      "properties": {
        "allCount": {
          "type": "integer",
          "format": "int32"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "maxPages": {
          "type": "integer",
          "format": "int32"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ApiUserDtoOut"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "description": "Copy 'Bearer ' + valid api key into field",
      "name": "Authorization",
      "in": "header"
    }
  }
}